LayoutTests/fast/scrolling/latching/nested-cross-axis-latch-expiration-expected.txt

PASS innerOverflowScrollEventCount > 0 is true
PASS outerOverflowScrollEventCount > 0 is true
PASS windowScrollEventCount is 0
PASS successfullyParsed is true

TEST COMPLETE
